ui_init: Move update_invite_user_option to sidebar_ui.
